+++ src/main/java/org/servantscode/donation/rest/PledgeSvc.java
@@ -4,6 +4,7 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.omg.CosNaming.NamingContextPackage.NotFound;
 import org.servantscode.commons.EnumUtils;
+import org.servantscode.commons.rest.SCServiceBase;
 import org.servantscode.donation.Pledge;
 import org.servantscode.donation.db.PledgeDB;
 
@@ -17,11 +18,12 @@ import java.util.stream.Stream;
 import static javax.ws.rs.core.MediaType.APPLICATION_JSON;
 
 @Path("/pledge")
-public class PledgeSvc {
+public class PledgeSvc extends SCServiceBase {
     private static final Logger LOG = LogManager.getLogger(PledgeSvc.class);
 
     @GET @Path("/family/{familyId}") @Produces(APPLICATION_JSON)
     public Pledge getFamilyPledges(@PathParam("familyId") int familyId) {
+        verifyUserAccess("pledge.read");
         try {
             return new PledgeDB().getActivePledge(familyId);
         } catch(Throwable t) {
@@ -32,6 +34,7 @@ public class PledgeSvc {
 
     @GET @Path("/family/{familyId}/history") @Produces(APPLICATION_JSON)
     public List<Pledge> getFamilyPledgeHistory(@PathParam("familyId") int familyId) {
+        verifyUserAccess("pledge.read");
         try {
             return new PledgeDB().getFamilyPledges(familyId);
         } catch(Throwable t) {
@@ -42,6 +45,7 @@ public class PledgeSvc {
 
     @POST @Consumes(APPLICATION_JSON) @Produces(APPLICATION_JSON)
     public Pledge createPledge(Pledge pledge) {
+        verifyUserAccess("pledge.create");
         try {
             return new PledgeDB().createPledge(pledge);
         } catch(Throwable t) {
@@ -53,6 +57,7 @@ public class PledgeSvc {
     @PUT @Path("/{pledgeId}") @Consumes(APPLICATION_JSON) @Produces(APPLICATION_JSON)
     public Pledge updatePledge(@PathParam("pledgeId") int pledgeId,
                                   Pledge pledge) {
+        verifyUserAccess("pledge.update");
         try {
             if(pledge.getId() != pledgeId)
                 throw new BadRequestException();
@@ -69,6 +74,7 @@ public class PledgeSvc {
 
     @DELETE @Path("/{pledgeId}") @Produces(APPLICATION_JSON)
     public void deletePledge(@PathParam("pledgeId") int pledgeId) {
+        verifyUserAccess("pledge.delete");
         try {
             if(!new PledgeDB().deletePledge(pledgeId))
                 throw new NotFoundException();
@@ -79,9 +85,7 @@ public class PledgeSvc {
     }
 
     @GET @Path("/types") @Produces(APPLICATION_JSON)
-    public List<String> getPledgeTypes() {
-        return EnumUtils.listValues(Pledge.PledgeType.class);
-    }
+    public List<String> getPledgeTypes() { return EnumUtils.listValues(Pledge.PledgeType.class); }
 
     @GET @Path("/freqs") @Produces(APPLICATION_JSON)
     public List<String> getPledgeFrequencies() {
